chore: #94 rename test-blocks to test-block
